pyproject.toml
[build-system]
requires = ['poetry-core>=1.0.0']
build-backend = 'poetry.core.masonry.api'

[tool.poetry]
name = 'fuckdl'
version = '0.1.0'
description = 'Playready and Widevine DRM downloader and decrypter'
authors = []

[tool.poetry.dependencies]
subby = {path = "./scripts/subby", develop = true}
pyplayready = {path = "./scripts/pyplayready-0.6.0", develop = true}	
python = "^3.9"
appdirs = "^1.4.4"
aiohttp = {version = "^3.9.1", optional = true}
beautifulsoup4 = "~4.11.2"
brotli = "^1.2.0"
click = "^8.1.3"
cffi = "^2.0.0"
coloredlogs = "^15.0"
construct = "2.8.8"
crccheck = "^1.0"
cryptography = "^43.0.3"
ecpy = "^1.2.5"
httpx = "^0.23.0"
isodate = "^0.6.1"
jsonpickle = "^2.0.0"
langcodes = { extras = ["data"], version = "^3.4.0" }
language-data = "1.2.0"
lxml = "^6.0.0"
m3u8 = "^0.9.0"
marisa-trie = "^1.1.0"
pproxy = "^2.7.7"
protobuf = "^3.13.0"
pycaption = "^2.1.1"
pycryptodome = "^3.21.0"
pycryptodomex = "^3.4.3"
pycountry = "^24.6.1"
pyhulu = "^1.1.2"
pymediainfo = "^5.0.3"
PyMySQL = { extras = ["rsa"], version = "^1.0.2" }
pysubs2 = "^1.6.1"
PyYAML = "^6.0.1"
requests = { extras = ["socks"], version = "2.28.0" }
ruamel-yaml = "^0.18.15"
setuptools=	"80.6.0"			
tldextract = "^3.1.0"
toml = "^0.10.2"
tqdm = "^4.67.1"
Unidecode = "^1.2.0"
urllib3 = "1.26.20"
validators = "^0.18.2"
websocket-client = "^1.1.0"
xmltodict = "^0.14.2"
yt-dlp = "^2022.11.11"
ftfy = "^6.3.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^3.8.4"
isort = "^5.9.2"
pyinstaller = "^4.4"
pyarmor = "^8.5.0"

[tool.poetry.scripts]
vt = 'fuckdl.fuckdl:main'
fuckdl = 'fuckdl.fuckdl:main'

[tool.isort]
line_length = 120
classes = ['CTV', 'FPS', 'IO', 'iTunes', 'MP4', 'TVNOW']
extend_skip = ['fuckdl/vendor']
